+++ primitives/src/lib.rs
@@ -1,0 +1,72 @@
+#![cfg_attr(not(feature = "std"), no_std)]
+
+#![allow(clippy::identity_op)]
+
+pub use constants::*;
+pub use opaque::*;
+pub use types::*;
+
+/// Common types of parachains.
+mod types {
+    use::sp_runtime::{
+        traits::{IdentifyAccount, Verify},
+        MultiSignature,
+    };
+    /// Alias to 512-bit hash when used in the context of a transaction signature on the chain.
+    pub type Signature = MultiSignature;
+
+    /// Some way of identifying an account on the chain. We intentionally make it equivalent
+    /// to the public key of our transaction signing scheme.
+    pub type AccountId = <<Signature as Verify>::Signer as IdentifyAccount>::AccountId;
+
+    /// Balance of an account.
+    pub type Balance = u128;
+
+    /// Index of a transaction in the chain.
+    pub type Index = u32;
+
+    /// A hash of some data used by the chain.
+    pub type Hash = sp_core::H256;
+
+    /// An index to a block.
+    pub type BlockNumber = u32;
+}
+
+/// Common constants of parachains.
+mod constants {
+	use super::types::BlockNumber;
+
+    /// This determines the average expected block time that we are targeting.
+    /// Blocks will be produced at a minimum duration defined by `SLOT_DURATION`.
+    /// `SLOT_DURATION` is picked up by `pallet_timestamp` which is in turn picked
+    /// up by `pallet_aura` to implement `fn slot_duration()`.
+    ///
+    /// Change this to adjust the block time.
+    pub const MILLISECS_PER_BLOCK: u64 = 12000;
+
+    // NOTE: Currently it is not possible to change the slot duration after the chain has started.
+    //       Attempting to do so will brick block production.
+    pub const SLOT_DURATION: u64 = MILLISECS_PER_BLOCK;
+
+    // Time is measured by number of blocks.
+    pub const MINUTES: BlockNumber = 60_000 / (MILLISECS_PER_BLOCK as BlockNumber);
+    pub const HOURS: BlockNumber = MINUTES * 60;
+    pub const DAYS: BlockNumber = HOURS * 24;
+}
+
+/// Opaque types. These are used by the CLI to instantiate machinery that don't need to know
+/// the specifics of the runtime. They can then be made to be agnostic over specific formats
+/// of data like extrinsics, allowing for them to continue syncing the network through upgrades
+/// to even the core data structures.
+pub mod opaque {
+	use super::*;
+	use sp_runtime::{generic, traits::BlakeTwo256};
+
+	pub use sp_runtime::OpaqueExtrinsic as UncheckedExtrinsic;
+	/// Opaque block header type.
+	pub type Header = generic::Header<BlockNumber, BlakeTwo256>;
+	/// Opaque block type.
+	pub type Block = generic::Block<Header, UncheckedExtrinsic>;
+	/// Opaque block identifier type.
+	pub type BlockId = generic::BlockId<Block>;
+}
